requirements.txt
tiktoken
networkx
pyautogen==0.2.35
pandas
sqlalchemy
gymnasium
numpy
rich
gymnasium
pyautogen
azure-identity
azure-monitor-query
mysql-connector-python
python-Levenshtein
matplotlib
azure-ai-ml
azure-ai-inference
bs4
yaml